+++ apps/deploy-web/src/context/CertificateProvider/index.ts
@@ -1,1 +1,2 @@
-export { useCertificate, CertificateProvider } from "./CertificateProviderContext";
+export { useCertificate, CertificateProvider, DEPENDENCIES } from "./CertificateProviderContext";
+export type { LocalCert } from "./CertificateProviderContext";