+++ apps/deploy-web/src/utils/proto/index.ts
@@ -1,8 +1,7 @@
 import * as v1beta3 from "@akashnetwork/akash-api/v1beta3";
 import * as v1beta4 from "@akashnetwork/akash-api/v1beta4";
+import type { NetworkId } from "@akashnetwork/akashjs/build/types/network";
 import { MAINNET_ID, SANDBOX_ID, TESTNET_ID } from "@akashnetwork/network-store";
-
-import networkStore from "@src/store/networkStore";
 
 const commonTypes = { ...v1beta3, ...v1beta4 };
 const mainnetTypes = commonTypes;
@@ -10,8 +9,8 @@ const sandboxTypes = commonTypes;
 
 export let protoTypes: typeof commonTypes | typeof mainnetTypes | typeof sandboxTypes;
 
-export function initProtoTypes() {
-  switch (networkStore.selectedNetworkId) {
+export function initProtoTypes(config: ProtoConfig) {
+  switch (config.networkId) {
     case MAINNET_ID:
     case TESTNET_ID:
       protoTypes = mainnetTypes;
@@ -25,4 +24,8 @@ export function initProtoTypes() {
       protoTypes = mainnetTypes;
       break;
   }
+}
+
+export interface ProtoConfig {
+  networkId: NetworkId;
 }